fix some environment setting issue
