+++ .gitignore
@@ -139,3 +139,5 @@ notebooks/playground.ipynb
 notebooks/check_dataset.ipynb
 data/swe-bench/swe-bench.json
 data/swe-bench/ToyExamples_dataset.json
+patch.diff
+testbed/
