+++ docker/swe.Dockerfile
@@ -15,6 +15,8 @@ RUN apt-get update && \
 RUN wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-aarch64.sh -O miniconda.sh
 RUN bash miniconda.sh -b -u -p ./miniconda3
 RUN rm miniconda.sh
+# support python=3.6 download
+RUN /miniconda3/bin/conda config --add channels conda-forge
 RUN export PATH=/miniconda3/bin:$PATH
 
 # for Github: https://stackoverflow.com/a/67390274
