+++ scripts/expr_multi_turn.sh
@@ -19,17 +19,17 @@
 # - chat
 
 # Bash Call
-python -m experiments.eval_n_turn \
-    --data_path ./data/nl2bash/nl2bash_fs_1.json \
-    --dialogue_limit 7 \
-    --env bash \
-    --image_name intercode-nl2bash \
-    --log_dir logs/experiments \
-    --max_turns 10 \
-    --policy chat \
-    --template v2 \
-    --model gpt-3.5-turbo \
-    --verbose
+# python -m experiments.eval_n_turn \
+#     --data_path ./data/nl2bash/nl2bash_fs_1.json \
+#     --dialogue_limit 7 \
+#     --env bash \
+#     --image_name intercode-nl2bash \
+#     --log_dir logs/experiments \
+#     --max_turns 10 \
+#     --policy chat \
+#     --template v2 \
+#     --model gpt-3.5-turbo \
+#     --verbose
 
 # SQL Call
 # python -m experiments.eval_n_turn \
@@ -70,3 +70,16 @@ python -m experiments.eval_n_turn \
 #     --template ctf \
 #     --model gpt-4 \
 #     --verbose
+
+SWE Call
+python -m experiments.eval_n_turn \
+    --data_path ./data/swe-bench/ic_swe_bench_dev.json \
+    --dialogue_limit 20 \
+    --env bash \
+    --image_name intercode-swe \
+    --log_dir logs/experiments \
+    --max_turns 10 \
+    --policy chat \
+    --template v2 \
+    --model gpt-3.5-turbo \
+    --verbose
